sync core library, prepare typespec-java 0.41.0 and autorest.java 4.1.63 (#3298)
* update core

* check-in generated sample code for premium

* try pin autorest version

* fix

* also fix autorest pipeline

* try failOnStderr

* Revert "try failOnStderr"

This reverts commit da7a11b7edb977a7cf91f0d0131fa6dd91811365.

* core on main

* prepare typespec-java 0.41.0

* prepare autorest.java 4.1.63

* sync main

* sync tests

* audit fix

* Update core submodule to latest upstream main

* Update Node.js packages for latest dependencies

+++ eng/sdk/sync_sdk.py
@@ -172,7 +172,9 @@ def update_sdks():
             cmd = ["git", "checkout", "src/test"]
             subprocess.check_call(cmd, cwd=module_path)
 
-        if not generated_samples_exists:
+        # For ARM module, we want to keep the generated samples code.
+        # For data-plane, if the generated samples/test code is not there before generation, we will delete the generated code after generation, to avoid unnecessary code check-in.
+        if not generated_samples_exists and not arm_module:
             shutil.rmtree(generated_samples_path, ignore_errors=True)
         if not generated_test_exists:
             shutil.rmtree(generated_test_path, ignore_errors=True)
